interlinking works now

+++ setup/htmlize-file.sh
@@ -4,30 +4,38 @@
 # shenanigans. if it seems to be failing and you can't see why because of this
 # setup, add '-nw' before the '-l' argument and run it in a terminal.
 
-# note that this won't work if you require interactivity with your emacs init
-# when it sets up (for typing in a password or something). if you cannot disable
-# that, you'll have to either 1) add '--quick' before '-l', and live without
-# your emacs installation being loaded, or 2) perform the interaction in '-nw'
-# mode
+# note that this will be annoying if you require interactivity with your emacs
+# init when it sets up (for typing in a password or something). if you cannot
+# disable that, you'll have to either add '--quick' before '-l', and live
+# without your emacs installation being loaded, or have some custom setup
 
 WORKING_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 tmpfile="$2"
 output_file="$WORKING_DIR/output-file"
 
-Xvfb "$1" -screen 1 800x600x24 &
-xvfb_pid="$!"
+display_num="$1"
+
+if hash 2>/dev/null; then
+  Xvfb "$display_num" -screen 1 800x600x24 &
+  xvfb_pid="$!"
+fi
+
 rm -f "$output_file"
 touch "$output_file"
+# tail is being used as a crude form of ipc since it's otherwise impossible to
+# communicate with a running emacs unless that emacs is running in batch mode
 tail -f "$output_file" &
 tail_pid="$!"
 
 function finish {
-  kill "$xvfb_pid"
+  [ "$xvfb_pid" != "" ] && kill "$xvfb_pid"
   kill "$tail_pid"
   rm "$tmpfile"
   rm "$output_file"
 }
 
 trap finish EXIT
-TERM="xterm" DISPLAY="$1" emacs -l "$WORKING_DIR/htmlize-this-file.el"
+
+# should work if in non-graphical environment too
+TERM="xterm" DISPLAY="$display_num" emacs -l "$WORKING_DIR/htmlize-this-file.el"
